modernizing tests
